+++ lib/dsapp_ghc.py
@@ -1697,3 +1697,4 @@ def ghc_verifyServerDate():
 
 	time2 = time.time()
 	logger.debug("Operation took %0.3f ms" % ((time2 - time1) * 1000))
+